# modulo_administracion_vehiculos/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Vehículos",
        "description": "Documentación de la API del sistema de administración de vehículos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/mantenimientos": {
            "get": {
                "tags": [
                    "Mantenimiento"
                ],
                "summary": "Lista todos los mantenimientos",
                "description": "Obtiene todos los registros de mantenimiento con su vehículo asociado",
                "operationId": "98eb8bd3b5c15a4078751d0c2f9d12ad",
                "responses": {
                    "200": {
                        "description": "Lista de mantenimientos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mantenimiento"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Mantenimiento"
                ],
                "summary": "Crea un nuevo mantenimiento",
                "description": "Crea un registro de mantenimiento asociado a un vehículo existente",
                "operationId": "2e7dbf32fd0b9bd23aa7c42ce7a2d530",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehiculo_id",
                                    "descripcion",
                                    "fecha"
                                ],
                                "properties": {
                                    "vehiculo_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Cambio de aceite"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mantenimiento creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mantenimiento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Vehículo no encontrado o datos inválidos"
                    }
                }
            }
        },
        "/api/mantenimientos/{id}": {
            "get": {
                "tags": [
                    "Mantenimiento"
                ],
                "summary": "Obtiene un mantenimiento por ID",
                "description": "Devuelve un registro de mantenimiento específico con su vehículo asociado",
                "operationId": "f1eccfea98c96e09c78b65db74a6c46e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del mantenimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del mantenimiento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mantenimiento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mantenimiento no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mantenimiento"
                ],
                "summary": "Actualiza un mantenimiento existente",
                "description": "Actualiza un registro de mantenimiento por ID",
                "operationId": "7b682cf6eb6fd45d9f93d8d5132d63f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del mantenimiento a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vehiculo_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Revisión de frenos"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-11"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mantenimiento actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mantenimiento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mantenimiento no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Mantenimiento"
                ],
                "summary": "Elimina un mantenimiento",
                "description": "Elimina un registro de mantenimiento por ID",
                "operationId": "d624e7b533b922ada73de58f62ed764d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del mantenimiento a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Mantenimiento eliminado correctamente"
                    },
                    "404": {
                        "description": "Mantenimiento no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Mantenimiento": {
                "required": [
                    "vehiculo_id",
                    "fecha_inicio",
                    "fecha_fin",
                    "descripción",
                    "costo"
                ],
                "properties": {
                    "id_mantenimiento": {
                        "type": "integer",
                        "example": 1
                    },
                    "vehiculo_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "fecha_inicio": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-01"
                    },
                    "fecha_fin": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-03"
                    },
                    "descripción": {
                        "type": "string",
                        "example": "Cambio de aceite y filtros"
                    },
                    "costo": {
                        "type": "number",
                        "format": "float",
                        "example": 120.5
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}